Reordering of the currentStatus struct
Reordered to match comms and cancomms serial "A" command.
Also changing the types of most of it's members to a fixed size.

+++ speeduino/corrections.ino
@@ -106,7 +106,7 @@ static inline byte correctionWUE()
   else
   {
     BIT_SET(currentStatus.engine, BIT_ENGINE_WARMUP);
-    WUEValue = table2D_getValue(&WUETable, currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET);
+    WUEValue = table2D_getValue(&WUETable, (int)currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET);
   }
 
   return WUEValue;
@@ -122,7 +122,7 @@ static inline byte correctionCranking()
   //if ( BIT_CHECK(currentStatus.engine, BIT_ENGINE_CRANK) ) { crankingValue = 100 + configPage2.crankingPct; }
   if ( BIT_CHECK(currentStatus.engine, BIT_ENGINE_CRANK) )
   {
-    crankingValue = table2D_getValue(&crankingEnrichTable, currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET);
+    crankingValue = table2D_getValue(&crankingEnrichTable, (int)currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET);
   }
   return crankingValue;
 }
@@ -138,10 +138,10 @@ static inline byte correctionASE()
   //Two checks are requiredL:
   //1) Is the negine run time less than the configured ase time
   //2) Make sure we're not still cranking
-  if ( (currentStatus.runSecs < (table2D_getValue(&ASECountTable, currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET))) && !(BIT_CHECK(currentStatus.engine, BIT_ENGINE_CRANK)) )
+  if ( (currentStatus.runSecs < (table2D_getValue(&ASECountTable, (int)currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET))) && !(BIT_CHECK(currentStatus.engine, BIT_ENGINE_CRANK)) )
   {
     BIT_SET(currentStatus.engine, BIT_ENGINE_ASE); //Mark ASE as active.
-    ASEValue = 100 + table2D_getValue(&ASETable, currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET);
+    ASEValue = 100 + table2D_getValue(&ASETable, (int)currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET);
   }
   else
   {
@@ -501,7 +501,7 @@ static inline int8_t correctionCLTadvance(int8_t advance)
 {
   int8_t ignCLTValue = advance;
   //Adjust the advance based on CLT.
-  int8_t advanceCLTadjust = (int16_t)(table2D_getValue(&CLTAdvanceTable, currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET)) - 15;
+  int8_t advanceCLTadjust = (int16_t)(table2D_getValue(&CLTAdvanceTable, (int)currentStatus.coolant + CALIBRATION_TEMPERATURE_OFFSET)) - 15;
   ignCLTValue = (advance + advanceCLTadjust);
   
   return ignCLTValue;
